precision is now initialised as -1 to check is it is not called

+++ include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/14 12:00:17 by pnguyen-          #+#    #+#             */
-/*   Updated: 2023/11/14 17:40:57 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2023/11/14 18:21:29 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define FT_PRINTF_H
 
 #include <stdarg.h>
-
+#include <sys/types.h>
 #include "ft_printfparser.h"
 
 typedef unsigned int uint;
@@ -22,6 +22,6 @@ typedef unsigned int uint;
 int		ft_printf(char const format[], ...);
 void	padding_space(unsigned int len);
 ssize_t	ft_printfchar(va_list ap, t_flags, uint width);
-ssize_t	ft_printfstr(va_list ap, t_flags, uint width, uint precision);
+ssize_t	ft_printfstr(va_list ap, t_flags, uint width, ssize_t precision);
 
 #endif
